Policies ok

+++ Alumni/app/Http/Controllers/JobController.php
@@ -53,9 +53,12 @@ class JobController extends Controller
      */
     public function update(Request $request, $id)
     {
-        // update a job
+        
         // toDo: add validator
         $job = Job::find($id);
+        //Check Policy first
+        $this->authorize('update', $job);
+        // update a job
         $job -> update($request->all());
         return $job;
 
@@ -69,7 +72,11 @@ class JobController extends Controller
      */
     public function destroy($id)
     {
-        // delete a job
+        //Get job id
+        $job = Job::find($id);
+        //check policy first
+        $this->authorize('delete', $job);
+        //DELETE a job
         return Job::destroy($id);
     }
 }
